MDEV-24519: Server crashes in Charset::set_charset upon SELECT
The query causing the issue here has implicit grouping for we
have to produce one row with special values for the aggregates
(depending on each aggregate function), and NULL values for all
non-aggregate fields.

The subselect item where implicit grouping was being done, null_value
for the subselect item was not being set for the case when the implicit
grouping produces NULL values for the items in the select
list of the subquery. This which was leading to the crash.

The fix would be to set the null_value when all the values for the row
column have NULL values.