MDEV-24949: Enable idle flushing (after MDEV-23855 disabled it)
Thanks to Krunal Bauskar for noticing this regression and providing
a suggested fix that this one is based on.

After MDEV-23855, the page cleaner thread will stop flushing if
dirty_pct < innodb_max_dirty_pages_pct_lwm.

The page cleaner used to flush all pages when the server is idle
(no write transactions have been committed recently). That is useful,
because it will limit the maximum checkpoint age and thus reduce
the likelihood of log_free_check() stalls during a write burst.

With this change, the page cleaner will avoid entering untimed sleep
when the server is not executing write transactions, and it will
keep writing up to innodb_io_capacity pages per second until the
buffer pool is clean or some write activity is recognized.